Refactor: 마이 대시보드에서 쓰는 svg파일 네이밍 수정
