+++ public/images/mocks/hero.png
